+++ likeminds-chat-reactnative-integration/ChatSX/context/Chat.tsx
@@ -2,12 +2,15 @@ import React from "react";
 import { ChatroomContextProvider } from "./ChatroomContext";
 import { MessageListContextProvider } from "./MessageListContext";
 import { ExploreFeedContextProvider } from "./ExploreFeedContext";
+import { InputBoxContextProvider } from "./InputBoxContext";
 
 function Chat({ children }: any) {
   return (
     <ChatroomContextProvider>
       <MessageListContextProvider>
-        <ExploreFeedContextProvider>{children}</ExploreFeedContextProvider>
+        <InputBoxContextProvider>
+          <ExploreFeedContextProvider>{children}</ExploreFeedContextProvider>
+        </InputBoxContextProvider>
       </MessageListContextProvider>
     </ChatroomContextProvider>
   );
