+++ likeminds-chat-reactnative-integration/ChatSX/context/CustomisableMethodsContext.tsx
@@ -19,6 +19,10 @@ export interface CustomisableMethodsContextProps {
   joinSecretChatroomProp?: () => void;
   showJoinAlertProp?: () => void;
   showRejectAlertProp?: () => void;
+  handleGalleryProp?: () => void;
+  handleCameraProp?: () => void;
+  handleDocProp?: () => void;
+  onEditProp?: () => void;
 }
 
 export interface CustomisableMethodsContext {}
@@ -56,6 +60,10 @@ export const CustomisableMethodsContextProvider = ({
   joinSecretChatroomProp,
   showJoinAlertProp,
   showRejectAlertProp,
+  handleGalleryProp,
+  handleCameraProp,
+  handleDocProp,
+  onEditProp,
 }: CustomisableMethodsContextProps) => {
   const contextValues: CustomisableMethodsContext = {
     setChatroomTopicProp,
@@ -75,6 +83,10 @@ export const CustomisableMethodsContextProvider = ({
     joinSecretChatroomProp,
     showJoinAlertProp,
     showRejectAlertProp,
+    handleGalleryProp,
+    handleCameraProp,
+    handleDocProp,
+    onEditProp,
   };
 
   return (
